requirements.txt
requests>=2.28
boto3>=1.26
python-dotenv>=1.0
pytest>=7.0